Make level-based filtering fundamental, not just the first pipeline stage
This is a smidge less general, but it's faster for messages that are
going to be filtered out: we don't have to create a Record object,
call the time() function, iterate over the pipeline. Just compare
levels and get out fast.

Also, it emphasizes that filtering based on log level is a fundamental
feature of lolog, not just something done by a pipeline stage.